feat(queue): add DLQ backend with retries and idempotency
